Revert all non-refactors
