+++ sql/core/src/test/scala/org/apache/spark/sql/JoinSuite.scala
@@ -27,7 +27,7 @@ import org.apache.spark.internal.config.SHUFFLE_SPILL_NUM_ELEMENTS_FORCE_SPILL_T
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.analysis.UnresolvedRelation
 import org.apache.spark.sql.catalyst.expressions.{Ascending, GenericRow, SortOrder}
-import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight, JoinSelectionHelper}
+import org.apache.spark.sql.catalyst.optimizer.{BuildLeft, BuildRight}
 import org.apache.spark.sql.catalyst.plans.logical.{Filter, HintInfo, Join, JoinHint, NO_BROADCAST_AND_REPLICATION}
 import org.apache.spark.sql.execution.{BinaryExecNode, FilterExec, ProjectExec, SortExec, SparkPlan, WholeStageCodegenExec}
 import org.apache.spark.sql.execution.adaptive.AdaptiveSparkPlanHelper
@@ -40,8 +40,7 @@ import org.apache.spark.sql.types.StructType
 import org.apache.spark.tags.SlowSQLTest
 
 @SlowSQLTest
-class JoinSuite extends QueryTest with SharedSparkSession with AdaptiveSparkPlanHelper
-  with JoinSelectionHelper {
+class JoinSuite extends QueryTest with SharedSparkSession with AdaptiveSparkPlanHelper {
   import testImplicits._
 
   setupTestData()
@@ -62,7 +61,6 @@ class JoinSuite extends QueryTest with SharedSparkSession with AdaptiveSparkPlan
     val sqlString = pair._1
     val c = pair._2
     val df = sql(sqlString)
-    val optimized = df.queryExecution.optimizedPlan
     val physical = df.queryExecution.sparkPlan
     val operators = physical.collect {
       case j: BroadcastHashJoinExec => j
@@ -76,10 +74,6 @@ class JoinSuite extends QueryTest with SharedSparkSession with AdaptiveSparkPlan
     if (operators.head.getClass != c) {
       fail(s"$sqlString expected operator: $c, but got ${operators.head}\n physical: \n$physical")
     }
-    assert(
-      canPlanAsBroadcastHashJoin(optimized.asInstanceOf[Join], conf) ===
-        operators.head.isInstanceOf[BroadcastHashJoinExec],
-      "canPlanAsBroadcastHashJoin not in sync with join selection codepath!")
     operators.head
   }
 
@@ -95,13 +89,11 @@ class JoinSuite extends QueryTest with SharedSparkSession with AdaptiveSparkPlan
       val planned = spark.sessionState.planner.JoinSelection(join)
       assert(planned.size == 1)
       assert(planned.head.isInstanceOf[CartesianProductExec])
-      assert(!canPlanAsBroadcastHashJoin(join, conf))
 
       val plannedWithHint = spark.sessionState.planner.JoinSelection(joinWithHint)
       assert(plannedWithHint.size == 1)
       assert(plannedWithHint.head.isInstanceOf[BroadcastNestedLoopJoinExec])
       assert(plannedWithHint.head.asInstanceOf[BroadcastNestedLoopJoinExec].buildSide == BuildLeft)
-      assert(!canPlanAsBroadcastHashJoin(joinWithHint, conf))
     }
   }
 
@@ -120,12 +112,10 @@ class JoinSuite extends QueryTest with SharedSparkSession with AdaptiveSparkPlan
     val planned = spark.sessionState.planner.JoinSelection(join)
     assert(planned.size == 1)
     assert(planned.head.isInstanceOf[BroadcastHashJoinExec])
-    assert(canPlanAsBroadcastHashJoin(join, conf))
 
     val plannedWithHint = spark.sessionState.planner.JoinSelection(joinWithHint)
     assert(plannedWithHint.size == 1)
     assert(plannedWithHint.head.isInstanceOf[SortMergeJoinExec])
-    assert(!canPlanAsBroadcastHashJoin(joinWithHint, conf))
   }
 
   test("NO_BROADCAST_AND_REPLICATION controls build side in BNLJ") {
@@ -141,13 +131,11 @@ class JoinSuite extends QueryTest with SharedSparkSession with AdaptiveSparkPlan
     assert(planned.size == 1)
     assert(planned.head.isInstanceOf[BroadcastNestedLoopJoinExec])
     assert(planned.head.asInstanceOf[BroadcastNestedLoopJoinExec].buildSide == BuildRight)
-    assert(!canPlanAsBroadcastHashJoin(join, conf))
 
     val plannedWithHint = spark.sessionState.planner.JoinSelection(joinWithHint)
     assert(plannedWithHint.size == 1)
     assert(plannedWithHint.head.isInstanceOf[BroadcastNestedLoopJoinExec])
     assert(plannedWithHint.head.asInstanceOf[BroadcastNestedLoopJoinExec].buildSide == BuildLeft)
-    assert(!canPlanAsBroadcastHashJoin(joinWithHint, conf))
   }
 
   test("join operator selection") {
@@ -202,13 +190,6 @@ class JoinSuite extends QueryTest with SharedSparkSession with AdaptiveSparkPlan
 //      ("SELECT * FROM arrayData JOIN complexData ON data = a", classOf[ShuffledHashJoin])
 //    ).foreach { case (query, joinClass) => assertJoin(query, joinClass) }
 //  }
-
-  test("broadcastable join with shuffle join hint") {
-    spark.sharedState.cacheManager.clearCache()
-    sql("CACHE TABLE testData")
-    assertJoin("SELECT /*+ SHUFFLE_HASH(testData) */ * FROM testData JOIN testData2 ON key = a",
-      classOf[ShuffledHashJoinExec])
-  }
 
   test("broadcasted hash join operator selection") {
     spark.sharedState.cacheManager.clearCache()
